requirements.txt
aiogram~=3.24.0
python-dotenv~=1.2.1
fastapi~=0.128.0
pytest~=7.4.0
httpx>=0.27.0

uvicorn
